feat: add commercial support and training page (#694)
## This PR

- adds a support and training page.
    - light/dark mode support
    - mobile friendly
- consolidate card styles
- added commercial support requirements

### Related Issues

Fixes #390 

### Notes

Direct link to the preview:
https://deploy-preview-694--openfeature.netlify.app/support-training

This is a continuation of @thisisobate
[PR](https://github.com/open-feature/openfeature.dev/pull/416). I
intended to push to that fork. Sorry, I didn't intend to hijack your PR.
If you'd prefer, I can close this, and we can keep working on the one
you started.

### Follow-up Tasks

Post in the OpenFeature Slack channel requesting vendors add themselves.

---------

Signed-off-by: thisisobate <obasiuche62@gmail.com>
Signed-off-by: Michael Beemer <beeme1mr@users.noreply.github.com>
Co-authored-by: thisisobate <obasiuche62@gmail.com>