docs: kubecon & tracking blog, tracking concept doc (#860)
A combined kubecon '24 and tracking blog post, as well as a basic
tracking concept doc.

Direct link to blog:
https://deploy-preview-860--openfeature.netlify.app/blog/kubecon-wrapped-and-tracking

Direct link to tracking concept doc:
https://deploy-preview-860--openfeature.netlify.app/docs/reference/concepts/tracking

If you want to add content to this, feel free to open a PR to this
branch and post it (and add yourself as author)! I've post-dated this to
Monday, so I won't be merging until then, earliest.

cc @roncohen @kozikow

---------

Signed-off-by: Todd Baert <todd.baert@dynatrace.com>