feat: Add the Vercel Flags SDK for Next.js + SvelteKit to the ecosystem page (#1042)
## This PR
- Adds Vercel's Flags SDK for Next.js and SvelteKit OpenFeature Adaptor
as an SDK under our ecosystem page
- Currently links out to their documentation, we could switch this to
our own documentation if that makes sense.
- I purposefully named this `Next.js Flags SDK` instead of just
`Next.js` if we decided to build our own SDK separate from this.
- Did not automatically pull in the parent providers as supported
provider as there is some uncertainty if all Node.js providers will work
with the Flags SDK. Vendors / projects will likely need to test their
support and list it as supported. (Also the listed `category` would have
to be `Server` for this to work).
- Labeled this SDK as a `Client` SDK even though it will be using the
OpenFeature Node.js SDK. My theory is that Next.js is thought of more as
a `Client` technology for building websites. We could potentially label
it as both `Client` and `Server`.

---------

Signed-off-by: Jonathan Norris <jonathan@taplytics.com>