Add Dropwizard OpenFeature Guide
Both SpringBoot as well as Dropwizard are examples of using OpenFeature in a
Java framework, adding a top level section for Java SDK with Spring Boot and
Dropwizard as examples.

Create documentation of how to integrate OpenFeature Java SDK in Dropwizard
using the InMemoryProvider either via the SDK only, or using the
dropwizard-openfeature module.

Signed-off-by: Philipp Fehre <philipp@fehre.co.uk>