Add Dropwizard OpenFeature Guide (#1192)
## This PR

Both Spring Boot and Dropwizard are examples of using OpenFeature in a
Java framework, adding a top level section for Java SDK with Spring Boot
and Dropwizard as examples.

~Create documentation of how to integrate OpenFeature Java SDK in
Dropwizard using the `InMemoryProvider` either via the SDK only, or
using the dropwizard-openfeature module.~

Create documentation on how to integrate OpenFeature Java SDK into
Dropwizard either by using the SDK directly, or the
https://github.com/sideshowcoder/dropwizard-openfeature module. In
either case using `flagd` as the provider of flags, with the shared
content from the Spring Boot tutorial.

---------

Signed-off-by: Philipp Fehre <philipp@fehre.co.uk>